feat(header): header 컴포넌트 분리
JIG-26
