+++ apps/web/src/components/basic/Header/_components/DesktopView/index.ts
@@ -1,0 +1,1 @@
+export { default as DesktopView } from './DesktopView';
