+++ apps/web/src/components/basic/Header/_components/MenuItem/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuItem';
