+++ apps/web/src/components/basic/Header/_components/MobileView/index.ts
@@ -1,0 +1,1 @@
+export { default as MobileView } from './MobileView';
