+++ apps/web/src/components/basic/Header/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './DesktopView';
+export * from './MenuItem';
+export * from './MobileView';
