+++ apps/web/src/styles/global.scss
@@ -11,8 +11,8 @@ html, body {
 }
 
 body {
-  font-family: var(--font-pretendard), "Noto Serif KR", "Malgun Gothic", "Helvetica Neue", sans-serif;
   margin: 0;
+  font-family: var(--font-pretendard), "Noto Serif KR", "Malgun Gothic", "Helvetica Neue", sans-serif;
 }
 
 
@@ -35,12 +35,12 @@ body {
   position: absolute;
   width: 1px;
   height: 1px;
-  padding: 0;
   margin: -1px;
+  padding: 0;
   overflow: hidden;
   clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
   border-width: 0;
+  white-space: nowrap;
 }
 
 
@@ -63,55 +63,55 @@ a {
 }
 
 button {
+  margin: 0;
   border: 0;
   outline: 0;
   background-color: transparent;
   cursor: pointer;
-  margin: 0;
 }
 
 .modal-backdrop {
-  position: absolute;
-  inset: 0;
-  background-color: rgb(0 0 0 / 70%);
   display: flex;
-  justify-content: center;
-  align-items: center;
+  position: absolute;
   z-index: 1000;
+  align-items: center;
+  justify-content: center;
+  background-color: rgb(0 0 0 / 70%);
+  inset: 0;
 }
 
 .modal {
+  display: flex;
+  position: relative;
+  align-items: center;
+  justify-content: center;
   width: 80%;
   max-width: 500px;
   height: auto;
   max-height: 500px;
+  padding: 20px;
   border: none;
   border-radius: 12px;
   background-color: white;
-  padding: 20px;
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
   font-size: 48px;
   font-weight: 500;
 }
 
 .close-button {
+  display: flex;
   position: absolute;
   top: 10px;
   right: 10px;
-  width: 48px;
-  height: 48px;
-  background-color: transparent;
-  border: none;
-  border-radius: 15px; /* Circular shape */
-  cursor: pointer;
-  display: flex;
   align-items: center;
   justify-content: center;
-  font-weight: 500;
+  width: 48px;
+  height: 48px;
+  border: none;
+  border-radius: 15px; /* Circular shape */
+  background-color: transparent;
   font-size: 24px; /* Adjust font size as needed */
+  font-weight: 500;
+  cursor: pointer;
 
   &:hover {
     background-color: #eee;
@@ -124,7 +124,7 @@ button {
 }
 
 .block-scrolling {
-  overflow: hidden;
   position: fixed;
   width: 100%;
+  overflow: hidden;
 }
